amitruck/pyproject.toml
[tool.poetry]
name = "amitruck"
version = "0.1.0"
description = ""
authors = ["jimmykamau <jimmykambiz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
djangorestframework = "^3.14.0"
dj-database-url = "^2.0.0"
mysqlclient = "^2.2.0"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
